requirements.txt
flask>=3.0.0
waitress>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pypdfium2>=4.0.0
google-generativeai>=0.7.0

# Uncomment when switching providers:
# anthropic>=0.30.0
# openai>=1.40.0
